Release 1.1 -Removed unnecessary methods from IFactoryStorage, and created IFactoryExpandedStorage interface to contain unessential methods -Created SideList to replace Map<Direction,?> generic map class, and optimize its use -Created StorageUtil class with static methods to help between platform storage transactions -Added FactoryAPIPlatform.getPlatformFactoryStorage method to "convert" the Mod Loader's block entity Storage to IPlatformHandlerApi, and facilitate its exposition -Added automatic(as a fallback) exposure of IPlatformHandlers for block entities that extend IFactoryStorage and correctly implement it in the getStorage method -Fixed BlockSide.convertToFacing method incorrect vertical plane orientation -Fixed block entity CompoundTag bug when loading progress(es) from older saves or corrupted saves -Renamed IFluidItem to IFluidHandlerItem and removed FluidStorageBuilder record -Created IEnergyStorageItem to facilitate Mod Loader's energy item implementation -and many others changes...

+++ common/src/main/java/wily/factoryapi/base/IPlatformFluidHandler.java
@@ -79,7 +79,7 @@ public interface IPlatformFluidHandler<T> extends ITagSerializable<CompoundTag>,
      * simulated) drained.
      */
     @NotNull
-    FluidStack drain(int maxDrain, boolean simulate);
+    FluidStack drain(long maxDrain, boolean simulate);
 
     default long getTotalSpace(){
         return Math.max(0, getMaxFluid() - getFluidStack().getAmount());
